types: Move `types::chain_header::ChainHeader` to `types::ChainHeader`
Also remove some enums that were causing name resolution issues.
