+++ crates/holochain/src/core/state/cascade/test.rs
@@ -9,11 +9,11 @@ use holochain_state::{
 };
 use holochain_types::{
     address::EntryAddress,
-    chain_header::ChainHeader,
     entry::Entry,
     header, observability,
     prelude::*,
     test_utils::{fake_agent_pubkey_1, fake_agent_pubkey_2, fake_header_hash},
+    ChainHeader,
 };
 use maplit::hashset;
 use mockall::*;
