+++ crates/holochain/src/core/state/source_chain.rs
@@ -6,9 +6,7 @@
 use holo_hash::*;
 use holochain_keystore::Signature;
 use holochain_state::{db::DbManager, error::DatabaseResult, prelude::Readable};
-use holochain_types::{
-    address::HeaderAddress, chain_header::ChainHeader, entry::Entry, prelude::*,
-};
+use holochain_types::{address::HeaderAddress, entry::Entry, prelude::*, ChainHeader};
 use shrinkwraprs::Shrinkwrap;
 
 pub use error::*;
