+++ crates/holochain/tests/cascade.rs
@@ -5,11 +5,11 @@ use holochain_2020::core::state::{
 };
 use holochain_state::{env::ReadManager, test_utils::test_cell_env};
 use holochain_types::{
-    chain_header::ChainHeader,
     entry::Entry,
     header,
     prelude::*,
     test_utils::{fake_agent_pubkey_1, fake_agent_pubkey_2, fake_header_hash},
+    ChainHeader,
 };
 
 fn fixtures() -> (
