+++ crates/types/src/address.rs
@@ -1,6 +1,6 @@
 //! wraps holo_hashes for the use of those hashes as storage addresses, either CAS or DHT
 
-use crate::{chain_header::ChainHeader, entry::Entry};
+use crate::{entry::Entry, ChainHeader};
 use holo_hash::*;
 use holochain_serialized_bytes::prelude::*;
